Fast check with λ-semantics - WIP
After so much internal back and forth, I realized λ-calculus readback
equivalence can be correctly performed by moving the λ-calculus
semantics to after the leap is completed. That is, the equality itself
uses the elegant algorithm, but it serves merely to identify leaps that
pass through paired ANNs. Then, when the leap is found, we must check
whether it is equivalent.

This is where the λ-calculus readback comes in. We traverse the leap
down to its deepmost wire, and then travel back towards root, except
with the λ-calculus readback in mind, applied to this specific path.
When doing so, whenever we pass through a port 2, this is an APP node,
so we call equality again on port 1, to check its argument. When we pass
through a port 1, this is a VAR port, so we need to determine the bruijn
index of that variable - which is just the remaining length of CON nodes
towards root. This algorithm is still incomplete, as we also need to
decay ANN nodes. A term is well typed when all ANN nodes can be
decayed, and the result of this process is the compiled term. This
should be easy to add.

I'm not sure this approach works, but it looks promising. After adding
the decay, I will proceed to test and think about this. If I'm
eventually convinced this is the way to go, I'll proceed to formalizing
and proving that ITT coherence corresponds to type-checking in affine
CoC. I can then work my way towards more powerful systems.

This addresses issue #1